feat(permissions): migrate checkResourceAccess off ctx.db to Supabase (closes #3867)
resourceInvites is Supabase-primary (in TABLE_MAP). The old checkResourceAccess
in _helpers/permissions.ts read it via ctx.db, which is stale for TABLE_MAP tables.
Migrated to an internalAction in _helpers/authAction.ts that reads from Supabase
via createSupabaseDb(), matching the established pattern for verifyOrgAccess and
checkPermission. Removed the stale ctx.db version and its now-unused requireUser import.
